Decoration Finished
BAKING LIGHT PROMBLEMS: tomorrow i will Delate all the Point Lights and i will put fewer to try and bake the maps (i tried to bake some times but this last one at 3:44 AM was taking to long, like 4 hours or so)

TO DO: Decorating the level that it's empty (Oscar Knows the one)
Do the Cutscene in game and Make the screenshots and Video.

Also Xin has to do somethings related to scripting i dont remember what due to the time.

You all did a great job ><